change gridfs chunkSize to Long to match Java driver
The Java driver sets a max chunk size of 3.5M but stores it in
mongo as a 64-bit number. We need to use Long so we don't throw
when we try to load the 64-bit number.

+++ mongo/src/main/scala/org/beaucatcher/mongo/gridfs/GridFSFile.scala
@@ -7,7 +7,7 @@ import org.beaucatcher.bson.Implicits._
 import org.joda.time._
 
 case class CreateOptions(filename : Option[String] = None, contentType : Option[String] = None,
-    chunkSize : Option[Int] = None, aliases : List[String] = Nil, metadata : BObject = BObject.empty) {
+    chunkSize : Option[Long] = None, aliases : List[String] = Nil, metadata : BObject = BObject.empty) {
 
     private[gridfs] def asBObjectWithNewId = {
         val b = BObject.newBuilder
@@ -50,7 +50,7 @@ class GridFSFile private[gridfs] (private[gridfs] val underlying : BObject) {
     /** get the file's length in bytes or throw NoSuchElementException */
     def length = underlying.getUnwrappedAs[Long]("length")
     /** get the file's chunk size or throw NoSuchElementException */
-    def chunkSize = underlying.getUnwrappedAs[Int]("chunkSize")
+    def chunkSize = underlying.getUnwrappedAs[Long]("chunkSize") // using "Long" is silly (Input/OutputStream don't) but Java driver stores as Int64
     /** get the file's uploadDate or throw NoSuchElementException */
     def uploadDate = underlying.getUnwrappedAs[DateTime]("uploadDate")
     /** get the file's aliases or throw NoSuchElementException */
